Revise GC statistics
As of version 1.8, GUILE uses lazy sweeping. This means that old
memory is freed ("swept") while new is being allocated. This has the
advantage that the sweep doesn't happen during a Stop-The-World phase.

There is a disadvantage, which is that a full view of the effects of
the mark phase are only available once the sweep finalizes, which
typically is when the next GC starts.

The previous code kept track of the number of cells collected
(scm_gc_collected), where 'collected' means: a cell transitions from
used (or new) state into ready for allocation (a with scm_tc_free_cell
tag).

This statistic was not just used for dumping (gc-stats) output. It was
also used for deciding when to grow the heap. In particular, if two GCs
in a row would have a low scm_gc_collected, this was a signal that
more heap was needed for storing cells.

This works in a normal situation, where the program consumes cells,
and a GC is triggered when there are no new cells to give
out.

There are 2 situation where this does not work:

1) There are situations where GUILE executes a full sweep before it
runs out of cells.  In particular, if a program allocates a lot of
malloc'd memory (strings, smobs etc.). GUILE will try a full sweep to
reduce malloc usage

In this case, the full sweep collects many cells, but afterwards,
statistics were reset, leading to low collect counts.

2) Any GC starts with a full sweep, so an explicit `(gc)` call will
sweep the current heap (leaving about many scm_tc_free_cells). When
calling `(gc)` twice, the second run will have a very low
scm_gc_collected count.

This had the following implications:

* Heap expansion was decided when basically by (scm_gc_collected <
  min_yield). When scm_gc_collected is too low, this leads GUILE to
  expand the heap when it was not necessary

* (gc-stats) tries to keep track of memory in-use (ie. either live or
  not yet freed) as the sweep went along. By underreporting the
  collected cells (in situation 1.), the in-use statistic grew too
  much, sometimes exceeding the heap size.

This commit fixes this by

* introducing the scm_t_sweep_statistics type, which distinguishes
  between:

  - unmarked cells (marked as garbage in last mark phase)
  - collected cells (cells transitioned from used to free)
  - swept cells (cells inspected to check if they should be collected)

  per cell-type.

* sweeps due to malloc usage affect the `collected` and `swept` metric, but not
  the `unmarked` metric.

* basing heap expansion decisions on the number of `unmarked` cells,
  and heap size (rather than a statistic that may accumulate errors),
  for the cell-type at hand.

* avoiding the use of accumulated statistics for these decisions.

An incomplete test is added as test.scm ; it needs more statistics to
measure current memory usage, and predict heap expansion sizes, but it
serves to highlight the problems with the previous GC strategy.

It would maybe better to a bit count over all the mark vectors at the
end of the marking phase, so statistics aren't delayed by a GC cycle,
but this adds extra work in the steady state case.

With this change, these are the stats for max RSS for lilypond at commit 
https://gitlab.com/hanwenn/lilypond/-/commit/6234db9f0acb7a1bd36c8601aa9d8e7202427c76

lilypond -f ps input/regression/d*.ly - 99M
lilypond -f ps input/regression/*.ly - 200M

removing GC between files,

lilypond -f ps input/regression/d*.ly - 107M
lilypond -f ps input/regression/*.ly - 225M

For the last test, GUILE 1.8 without this fix lets the heap balloon
to approximately 1 Gb.