gc: calculate min_yield statelessly
Before, the calculation of scm_t_cell_statistics was updated by adding
an increment. This makes it harder to reason about the memory increase
threshold.

This may create a small changes in memory use: before, the min_yield
for each freelist (counted in cells) was (min_yield_fraction *
total_heap_size), now it is computed separately for each cell-type.

In the past, it would have been possible for a large allocation of
4-SCM cells (large total SCM_HEAP_SIZE) to cause min_yield for 2-SCM
cells to become unrealistically high, forcing heap expansion for the
2-SCM cell types.